requirements_openpi.txt
absl-py==2.3.1
attrs==25.4.0
ClusterShell==1.9.3
contourpy==1.2.1
cycler==0.12.1
dbus-python==1.2.18
distro==1.5.0
docstring_parser==0.17.0
etils==1.5.2
eval_type_backport==0.2.2
evdev==1.9.2
fail2ban==1.1.0
fonttools==4.60.1
fsspec==2025.9.0
glfw==2.10.0
gpg==1.15.1
h5py==3.14.0
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==1.1.1
iotop==0.6
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
libcomps==0.1.18
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.8.4
mdurl==0.1.2
mpmath==1.3.0
mujoco==3.3.6
nftables==0.1
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
opencv-python==4.11.0.86
packaging==20.9
pandas==2.3.3
pillow==11.3.0
platformdirs==4.4.0
pluggy==0.13.1
py==1.10.0
pycairo==1.20.1
Pygments==2.19.2
PyGObject==3.40.1
pynput==1.8.1
PyOpenGL==3.1.10
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.9.0.post0
python-xlib==0.33
pytz==2025.2
PyYAML==5.4.1
referencing==0.36.2
rich==14.1.0
robosuite==1.4.1
rpds-py==0.27.1
rpm==4.16.1.3
scipy==1.13.1
selinux==3.6
sepolicy==3.6
setools==4.4.4
shtab==1.7.2
six==1.15.0
smmap==5.0.2
sympy==1.14.0
systemd-python==234
termcolor==3.1.0
toml==0.10.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.4.0
typeguard==4.4.4
typing_extensions==4.15.0
tyro==0.9.32
tzdata==2025.2
urllib3==1.26.20
zipp==3.23.0
